ROX-20479: rename package
