+++ pkg/serviceregistration/leader_test.go
@@ -1,4 +1,4 @@
-package leader
+package serviceregistration
 
 import (
 	"context"
